Conflicts:
	src/main/webapp/WEB-INF/jsp/ezEmail/mailSearchView.jsp

+++ src/main/java/egovframework/ezEKP/ezWebFolder/web/EzWebFolderGWController_y.java
@@ -393,8 +393,8 @@ public class EzWebFolderGWController_y extends EgovFileMngUtil {
 			}
 			
 			if (resmode.equals("copy")) {
-				double folderSize           = ezWebFolderAdminService.getFolderSize(folder.getFolderPath(), tenantId);
-				UserCapacityVO userCapacity = ezWebFolderAdminService.getUserCapacity(userInfo.getId(), userInfo.getPrimary(), userInfo.getTenantId());
+				double folderSize       = ezWebFolderAdminService.getFolderSize(folder.getFolderPath(), tenantId);
+				UserCapacityVO userCapacity = ezWebFolderAdminService.getCapacity(uppId, userInfo.getPrimary(), userInfo.getTenantId());
 				
 				double totalUsed = Double.parseDouble(userCapacity.getTotalUsed());
 				double totalCapa = Double.parseDouble(userCapacity.getTotalCapacity()) * 1073741824;
@@ -445,16 +445,22 @@ public class EzWebFolderGWController_y extends EgovFileMngUtil {
 		try {
 			common = mOptionService.commonInfoWeb(serverName, userId);
 			int tenantId 	= common.getTenantId();
-			String comId 	= common.getCompanyId();
-			String offset 	= common.getOffSet();
-			String timeUTC  = commonUtil.getDateStringInUTC(formatter.format(date), offset, true);
-			int result = service.deleteSubFldAFile(folderId, tenantId, comId, userId, timeUTC);
-			if (result == 1) {
-				jsonObj.put("status", "ok");
-				jsonObj.put("code", 0);
-			} else if (result == 2) {
+			
+			if (ezWebFolderService.canDelete(null, new String[] { folderId }, userId, tenantId)) {
+				String comId 	= common.getCompanyId();
+				String offset 	= common.getOffSet();
+				String timeUTC  = commonUtil.getDateStringInUTC(formatter.format(date), offset, true);
+				int result = service.deleteSubFldAFile(folderId, tenantId, comId, userId, timeUTC);
+				if (result == 1) {
+					jsonObj.put("status", "ok");
+					jsonObj.put("code", 0);
+				} else if (result == 2) {
+					jsonObj.put("status", "error");
+					jsonObj.put("code", 4);
+				}
+			} else {
 				jsonObj.put("status", "error");
-				jsonObj.put("code", 4);
+				jsonObj.put("code", 5);
 			}
 		} catch (Exception e) {
 			e.printStackTrace();
@@ -1059,7 +1065,6 @@ public class EzWebFolderGWController_y extends EgovFileMngUtil {
 			
 			LoginVO userInfo = commonUtil.getUserForGw(userId, serverName);
 			int tenantId = userInfo.getTenantId();
-			String primary = userInfo.getPrimary();
 			String realPath  = request.getServletContext().getRealPath("");
 			
 			if (!isWebfolderAdmin(userInfo)){
@@ -1086,7 +1091,7 @@ public class EzWebFolderGWController_y extends EgovFileMngUtil {
 				return result;
 			}
 			
-			UserCapacityVO userCapacity = ezWebFolderAdminService.getUserCapacity(userId, primary, userInfo.getTenantId());
+			UserCapacityVO userCapacity = ezWebFolderAdminService.getCapacity(folderId, userInfo.getPrimary(), userInfo.getTenantId());
 			LOGGER.debug("userCapacity!");
 			
 			double totalUsed = Double.parseDouble(userCapacity.getTotalUsed());
@@ -1149,9 +1154,6 @@ public class EzWebFolderGWController_y extends EgovFileMngUtil {
 	    	String userId = requestObject.get("userid") != null ? (String)requestObject.get("userid") : "";
 	    	String pw = requestObject.get("pw") != null ? (String)requestObject.get("pw") : "";
 	    	int tenantId = 0;
-
-	    	System.out.println(webfolderUtil.encryptAES("webfolder2"));
-	    	System.out.println(webfolderUtil.encryptAES("123qwe!!"));
 	    	
 	    	userId = webfolderUtil.decryptAES(userId);
 			pw = webfolderUtil.decryptAES(pw);
